// src/app/shared/models/AadharCard.ts
export class AadharCard  {

    constructor() {
        this.frontImage = "";
        this.backImage = "";
        
    }
    _id?: string;
    aadharNumber?: string;
    name?: string;
    dateOfBirth?: Date;
    address?: string;
    frontImage?: string;
    backImage?: string;
    createdAt?: Date;
    updatedAt?: Date;
}