// src/app/shared/models/Passport.ts
export class Passport  {
    /**
 *
 */
constructor() {
    this.frontImage = "";
    this.backImage = "";
    
}
    _id?: string;
    passportNumber?: string;
    givenName?: string;
    surname?: string;
    dateOfIssue?: Date;
    dateOfExpiry?: Date;
    frontImage?: string;
    backImage?: string;
    createdAt?: Date;
    updatedAt?: Date;
}